Madam/requirements.txt
streamlit
pandas
Pillow
plotly
plotly-express
numpy
nltk
wordcloud
matplotlib
pycountry
fuzzywuzzy